Move `BitmapUtil` & `FrameMetadata` to `util` package
