+++ app/src/main/java/org/simple/clinic/util/FrameMetadata.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.simple.clinic.widgets.qrcodescanner
+package org.simple.clinic.util
 
 /** Describing a frame info.  */
 class FrameMetadata private constructor(val width: Int, val height: Int, val rotation: Int) {
